1. Fixed the incorrect computation of availability values in function compute_availability() The "derate" dataframe was being generated with incorrect availability values (for pass3) in this function. There were two main reasons/issues that caused this incorrect calculation.
First, the merging with inputs_temporal results in each pass3 timepoint (except those on the last day) occurring in two horizons.
This happens because of the one day look-ahead feature that is typically used in pass3.
So the fix for this is to filter inputs_temporal results for only those rows where the values in spinup_or_lookahead column are 0.
This approach will work for pass2 timepoints as well.

Second, the subsequent merging with inputs_temporal_horizon_timepoints again results in each pass3 timepoint occurring in two horizons.
This is over and above the issue mentioned in the previous para - i.e. the effect gets compounded.
The fix for this is to merge with a correct timepoint-horizon map for pass2/pass3 (depending on the "-s2" command line parameter).
The correct timepoint-horizon map can be generated from the file that is passed via the "--timepoint_map" command line parameter.
A new function get_timepoint_horizon_map() has been implemented for the same.
Again, this approach will work for pass2 timeppoints as well.

2. Corrected the sequence of steps to adjust for availability values.
The correct sequence that has been now implemented in this check-in, is as follows.

A1. The values in the max_ vector are adjusted, so that they do not exceed the availability values in the derate vector.
A2. Then the values in the min_ vector are adjusted, so that they do not exceed the newly adjusted values in the max_ vector.
A new function compute_adjusted_min_max() has been implemented for the above adjustments.

B. Then, the existing adjust_mean_const() function is called to do its job.
This function essentially adjusts the values in the avg_ vector, such that their original sum remains the same,
but every value in avg_ ultimately satisfies:  min_  <=  avg_  <=  max_

C. Then, the existing function compute_adjusted_variables() is called so that the values in the avg_, min_ and max_ vectors
are now adjusted as per the availability values in the derate vector.
Essentially, we simply divide all values in each of these vectors by the corresponding availability value (if the latter is non-zero).